backend/requirements.txt
# CivicIQ Backend Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Supabase
supabase==2.10.0

# AI / HTTP
httpx==0.27.2
google-generativeai==0.8.3

# Auth / JWT
PyJWT==2.9.0

# Validation / Config
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic[email]==2.9.2

# Utility
python-multipart==0.0.17   # For file uploads
python-dotenv==1.0.1
